requirements.txt
python-dotenv>=1.0.1
yt-dlp>=2026.3.1
faster-whisper>=1.1.0
openai>=1.79.0
anthropic>=0.52.0
google-generativeai>=0.8.5
deep-translator>=1.11.4
